初步实现【路径追踪】，效果渣渣；渲染器支持【超级采样抗锯齿】；新增场景（CornellBall）

+++ Camera/Camera.cpp
@@ -10,7 +10,7 @@ void Camera::setResolution(int width, int height) {
     this->height = height;
 }
 
-Ray Camera::getRay(int x, int y) const {
+Ray Camera::getRay(float x, float y) const {
     auto delta = (up * (height / 2 - x) + right * (y - width / 2)) * (realw / width);
     if(orthographic)
         return Ray(pos + delta, target - pos);
